Catch spurious rosnode exceptions
Prior to this commit, `monitor.py` would frequently crash.
`rosnode.get_node_names()` and `rosnode.get_api_uri(...)` would both
occasionally raise a ROSNodeIOException that claimed they could not
communicate with master. However, since cpu_monitor had launched
successfully and there is no fancy networking between the node and
master, this is surprising. Inside the `rosnode` package, it looks like
this exception is raised no matter what kind of socket error is raised
when trying to communicate with master. Since we know there may be some
wierd socket-related issues between the Piksi and the TX2, it seems
likely that these socket errors are not caused by not having a route to
master, but rather by a different socket error. Unfortunately `rosnode`
does not disambiguate these in the exception it raises, so for now we
will just catch everything and continue accordingly.